Spotify-Clone
MyMusicApp is a Java-based Android application that offers a Spotify-inspired music streaming experience. Leveraging the robustness and widespread adoption of Java, it utilizes key technologies like Android Studio for development, Firebase for seamless backend services such as authentication and real-time database management, Retrofit for effortless integration with the Spotify API, and Glide for efficient image loading and caching. This combination ensures a feature-rich and high-performing music app tailored for the Android platform.